requirements.txt
asgiref==3.4.1
conventional-commit>=0.4.2
Django==4.0
django-DevOps==0.2.1
gunicorn==20.1.0
sqlparse==0.4.2
python-dotenv>=0.19.2
python-digitalocean>=1.17.0
